pyproject.toml
[tool.poetry]
name = "grocery-store"
version = "0.1.0"
description = ""
authors = ["Aleksandr <kashin152@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
black = "^25.1.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
flake8 = "^7.1.2"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
isort = "^6.0.1"
drf-yasg = "^1.21.10"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
